[Modify]<src/components/Modals/BugReport, ChangePw, Login. SignIn/...>Replace with extracted components
추출한 input, button 컴포넌트로 styled component 대신함
