+++ src/components/Modals/BugReport/BugReport.tsx
@@ -1,26 +1,16 @@
 import * as React from 'react';
-
-// import ModalCloseImg from '../../assets/icon/black-close-button.png';
-// import BugIconImg from '../../assets/icon/bug.png';
-
-import * as S from '../style';
+import * as S from './style';
+import ModalWrapper from '../ModalWrapper';
+import ModalButton from '../ModalButton';
 
 interface Props {}
 
 const BugReport: React.StatelessComponent<Props> = () => {
   return (
-    <>
-      <S.ModalNameWrapper>
-        <S.ModalName>
-          <S.BugIcon />
-          버그신고
-        </S.ModalName>
-        <S.ModalClosebtn />
-      </S.ModalNameWrapper>
+    <ModalWrapper modalName="버그신고">
       <S.BugreportInput placeholder="버그신고 내용을 입력해주세요" />
-
-      <S.ModalSubmitBtn>신고하기</S.ModalSubmitBtn>
-    </>
+      <ModalButton buttonName="신고하기" />
+    </ModalWrapper>
   );
 };
