+++ src/components/Modals/ChangePW/ChangePW.tsx
@@ -1,23 +1,21 @@
 import * as React from 'react';
-import ModalCloseImg from '../../assets/icon/black-close-button.png';
-import * as S from '../style';
+import * as S from './style';
+import ModalWrapper from '../ModalWrapper';
+import ModalButton from '../ModalButton';
+import ModalInput from '../ModalInput';
 
 interface Props {}
 
 const ChangePW: React.StatelessComponent<Props> = () => {
   return (
-    <>
-      <S.ModalNameWrapper>
-        <S.ModalName>비밀번호 변경</S.ModalName>
-        <S.ModalClosebtn src={ModalCloseImg} />
-      </S.ModalNameWrapper>
+    <ModalWrapper modalName="비밀번호 변경">
       <S.FindPWContentWrapper>
-        <S.ModalInput placeholder="현재 비밀번호" type="password" />
-        <S.ModalInput placeholder="새로운 비밀번호" type="password" />
-        <S.ModalInput placeholder="비밀번호 확인" type="password" />
+        <ModalInput placeholder="현재 비밀번호" inputType="password" />
+        <ModalInput placeholder="새로운 비밀번호" inputType="password" />
+        <ModalInput placeholder="비밀번호 확인" inputType="password" />
       </S.FindPWContentWrapper>
-      <S.ModalSubmitBtn>회원가입</S.ModalSubmitBtn>
-    </>
+      <ModalButton buttonName="비밀번호 변경" />
+    </ModalWrapper>
   );
 };
