+++ src/components/Modals/SignIn/SignIn.tsx
@@ -1,26 +1,22 @@
 import * as React from 'react';
-import CloseImg from '../../assets/icon/black-close-button.png';
+import CloseImg from '../../../assets/icon/black-close-button.png';
 import * as S from './style';
-
+import ModalWrapper from '../ModalWrapper';
+import ModalButton from '../ModalButton';
+import ModalInput from '../ModalInput';
 interface Props {}
 
 const SignIn: React.StatelessComponent<Props> = () => {
   return (
-    <>
-      <S.ModalNameWrapper>
-        <S.ModalName>회원가입</S.ModalName>
-        <S.ModalClosebtn src={CloseImg} />
-      </S.ModalNameWrapper>
-
+    <ModalWrapper modalName="회원가입">
       <S.SignInContentWrapper>
-        <S.ModalInput placeholder="확인코드" />
-        <S.ModalInput placeholder="아이디" />
-        <S.ModalInput placeholder="비밀번호" type="password" />
-        <S.ModalInput placeholder="비밀번호 확인" type="password" />
+        <ModalInput placeholder="확인코드" />
+        <ModalInput placeholder="아이디" />
+        <ModalInput placeholder="비밀번호" inputType="password" />
+        <ModalInput placeholder="비밀번호 확인" inputType="password" />
       </S.SignInContentWrapper>
-
-      <S.ModalSubmitBtn>회원가입</S.ModalSubmitBtn>
-    </>
+      <ModalButton buttonName="회원가입" />
+    </ModalWrapper>
   );
 };
